+++ backup_manager/__init__.py
@@ -1,6 +1,5 @@
 """
 backup_manager package
-
 Contains core backup management functionality.
 """